refactor(dialog): extract trip inspection stop list
Move trip stop list rendering into a dedicated component to reduce the size of the dialog component and keep its responsibilities focused on orchestration.
