+++ src/components/dialog/trip-inspection-stop-list.tsx
@@ -1,0 +1,302 @@
+import { DEFAULT_AGENCY_LANG, resolveAgencyLang } from '@/config/transit-defaults';
+import { LOW_CONTRAST_BADGE_MIN_RATIO } from '@/domain/transit/color-resolver/contrast-thresholds';
+import {
+  getContrastAdjustedRouteColors,
+  resolveRouteColors,
+} from '@/domain/transit/color-resolver/route-colors';
+import { getStopDisplayNames } from '@/domain/transit/get-stop-display-names';
+import { getTimetableEntryAttributes } from '@/domain/transit/timetable-entry-attributes';
+import { useInfoLevel } from '@/hooks/use-info-level';
+import { useThemeContrastAssessment } from '@/hooks/use-is-low-contrast-against-theme';
+import type { InfoLevel } from '@/types/app/settings';
+import type {
+  ContextualTimetableEntry,
+  SelectedTripSnapshot,
+  TripStopTime,
+} from '@/types/app/transit-composed';
+import { StopInfo } from '../stop-info';
+import { StopTimeDetailInfo } from '../stop-time-detail-info';
+import { StopTimeItem } from '../stop-time-item';
+import { StopTimeTimeInfo } from '../stop-time-time-info';
+
+interface TripInspectionStopListProps {
+  tripSnapshot: SelectedTripSnapshot;
+  renderedSnapshot: SelectedTripSnapshot | null;
+  selectedPatternStopIndex: number;
+  infoLevel: InfoLevel;
+  dataLangs: readonly string[];
+  now: Date;
+}
+
+interface TripInspectionStopRowProps {
+  tripStopTime: TripStopTime;
+  currentPatternStopIndex: number;
+  infoLevel: InfoLevel;
+  serviceDate: Date;
+  dataLangs: readonly string[];
+  now: Date;
+}
+
+interface TripInspectionPlaceholderRowProps {
+  stopIndex: number;
+  totalStops: number;
+  currentPatternStopIndex: number;
+  infoLevel: InfoLevel;
+}
+
+type RenderedTripStopRow =
+  | { kind: 'stop'; stop: TripStopTime; stopIndex: number; totalStops: number }
+  | { kind: 'placeholder'; stopIndex: number; totalStops: number };
+
+// Initial frame renders the selected stop and +/- 5 neighbors so
+// long trips paint quickly before the full list is restored.
+const INITIAL_TRIP_STOP_RENDER_PADDING = 5;
+
+function buildRenderedTripStopRows(stopTimes: readonly TripStopTime[]): RenderedTripStopRow[] {
+  if (stopTimes.length === 0) {
+    return [];
+  }
+
+  const totalStops = stopTimes[0]?.timetableEntry.patternPosition.totalStops ?? 0;
+  const stopByIndex = new Map(
+    stopTimes.map((stop) => [stop.timetableEntry.patternPosition.stopIndex, stop]),
+  );
+
+  return Array.from({ length: totalStops }, (_, stopIndex) => {
+    const stop = stopByIndex.get(stopIndex);
+    if (stop) {
+      return { kind: 'stop', stop, stopIndex, totalStops } satisfies RenderedTripStopRow;
+    }
+
+    return { kind: 'placeholder', stopIndex, totalStops } satisfies RenderedTripStopRow;
+  });
+}
+
+function TripInspectionStopRow({
+  tripStopTime,
+  currentPatternStopIndex,
+  infoLevel,
+  dataLangs,
+  serviceDate,
+  now,
+}: TripInspectionStopRowProps) {
+  const infoLevelFlag = useInfoLevel(infoLevel);
+  const stopMeta = tripStopTime.stopMeta;
+  const stopId = tripStopTime.stopMeta?.stop.stop_id || '(unknown-stop)';
+  const stopAttributes = getTimetableEntryAttributes(tripStopTime.timetableEntry);
+  const stopAgency = stopMeta?.agencies.find(
+    (agency) => agency.agency_id === tripStopTime.timetableEntry.routeDirection.route.agency_id,
+  );
+  const stopRoute = tripStopTime.timetableEntry.routeDirection.route;
+  const { routeColor } = resolveRouteColors(stopRoute, 'css-hex');
+  const routeColorAssessment = useThemeContrastAssessment(routeColor, LOW_CONTRAST_BADGE_MIN_RATIO);
+  const contrastAdjustedRouteColors = getContrastAdjustedRouteColors(
+    stopRoute,
+    routeColorAssessment.isLowContrast,
+    'css-hex',
+  );
+  const stopAgencyLangs = tripStopTime.stopMeta
+    ? resolveAgencyLang(tripStopTime.stopMeta.agencies, tripStopTime.stopMeta.stop.agency_id)
+    : DEFAULT_AGENCY_LANG;
+  const stopNames = tripStopTime.stopMeta
+    ? getStopDisplayNames(tripStopTime.stopMeta.stop, dataLangs, stopAgencyLangs)
+    : null;
+  const stopIndex = tripStopTime.timetableEntry.patternPosition.stopIndex;
+  const isCurrent = stopIndex === currentPatternStopIndex;
+  const isTerminalStop = tripStopTime.timetableEntry.patternPosition.isTerminal;
+  const isFirstStop = tripStopTime.timetableEntry.patternPosition.isOrigin;
+  const showArrivalTime = isTerminalStop || !isFirstStop;
+  const showDepartureTime = !isTerminalStop;
+
+  const contextualTimetableEntry: ContextualTimetableEntry = {
+    serviceDate: serviceDate,
+    ...tripStopTime.timetableEntry,
+  };
+
+  return (
+    <div
+      key={`${stopId}:${stopIndex}`}
+      data-trip-stop-index={stopIndex}
+      className={[
+        'rounded-md border px-3 py-2',
+        isCurrent ? 'border-primary bg-primary/5' : 'border-border bg-background',
+      ].join(' ')}
+    >
+      <div className="grid grid-cols-[minmax(0,1fr)_auto] items-start gap-3">
+        <div className="min-w-0">
+          {stopMeta ? (
+            <StopInfo
+              stop={stopMeta.stop}
+              agencies={stopMeta.agencies}
+              showAgencies={true}
+              routeTypes={tripStopTime.routeTypes}
+              showRouteTypes={true}
+              routes={stopMeta.routes}
+              showRoutes={true}
+              stats={stopMeta.stats}
+              geo={stopMeta.geo}
+              mapCenter={null}
+              infoLevel={infoLevel}
+              dataLangs={dataLangs}
+              agencyBadgeSize="sm"
+              routeBadgeSize="xs"
+            />
+          ) : (
+            <div className="flex min-w-0 flex-col gap-1">
+              {stopNames && stopNames.subNames.length > 0 && (
+                <div className="text-muted-foreground truncate text-xs">
+                  {stopNames.subNames.join(' / ')}
+                </div>
+              )}
+              <div className="flex items-center gap-2">
+                <span className="text-muted-foreground text-xs">#{stopIndex}</span>
+                <span className="truncate font-medium">{stopNames?.name || stopId}</span>
+              </div>
+              <div className="text-muted-foreground truncate text-xs">{stopId}</div>
+            </div>
+          )}
+        </div>
+        <StopTimeTimeInfo
+          arrivalMinutes={tripStopTime.timetableEntry.schedule.arrivalMinutes}
+          departureMinutes={tripStopTime.timetableEntry.schedule.departureMinutes}
+          serviceDate={serviceDate}
+          now={now}
+          size="md"
+          showArrivalTime={showArrivalTime}
+          showDepartureTime={showDepartureTime}
+          collapseArrivalWhenSameAsDeparture={true}
+          forceShowRelativeTime={true}
+          showVerbose={false}
+          textAppearance={{ color: contrastAdjustedRouteColors.color }}
+        />
+      </div>
+      <StopTimeDetailInfo
+        entry={tripStopTime.timetableEntry}
+        infoLevel={infoLevel}
+        dataLangs={dataLangs}
+        showRouteTypeIcon={false}
+        agency={stopAgency}
+        showAgency={false}
+        attributes={stopAttributes}
+      />
+
+      {infoLevelFlag.isVerboseEnabled && (
+        <>
+          <hr className="m-2" />
+          <StopTimeItem
+            entry={contextualTimetableEntry}
+            now={now}
+            showArrivalTime={showArrivalTime}
+            showDepartureTime={showDepartureTime}
+            collapseArrivalWhenSameAsDeparture={true}
+            forceShowRelativeTime={true}
+            showRouteTypeIcon={true}
+            infoLevel={infoLevel}
+            dataLangs={dataLangs}
+            agency={stopAgency}
+            showAgency={true}
+          />
+        </>
+      )}
+    </div>
+  );
+}
+
+function TripInspectionPlaceholderRow({
+  stopIndex,
+  totalStops,
+  currentPatternStopIndex,
+  infoLevel,
+}: TripInspectionPlaceholderRowProps) {
+  const infoLevelFlag = useInfoLevel(infoLevel);
+  const isCurrent = stopIndex === currentPatternStopIndex;
+
+  return (
+    <div
+      data-trip-stop-index={stopIndex}
+      className={[
+        'rounded-md border border-dashed px-3 py-2',
+        isCurrent ? 'border-primary bg-primary/5' : 'border-border bg-muted/20',
+      ].join(' ')}
+    >
+      <div className="flex min-w-0 flex-col gap-1">
+        <div className="flex items-center gap-2">
+          <span className="text-muted-foreground text-xs">#{stopIndex}</span>
+          <span className="text-muted-foreground truncate font-medium">Stop-time unavailable</span>
+        </div>
+        {infoLevelFlag.isVerboseEnabled && (
+          <div className="text-muted-foreground truncate text-xs">
+            Pattern stop {stopIndex + 1} / {totalStops}
+          </div>
+        )}
+      </div>
+    </div>
+  );
+}
+
+export function TripInspectionStopList({
+  tripSnapshot,
+  renderedSnapshot,
+  selectedPatternStopIndex,
+  infoLevel,
+  dataLangs,
+  now,
+}: TripInspectionStopListProps) {
+  const infoLevelFlag = useInfoLevel(infoLevel);
+  const renderedTripStopRows = buildRenderedTripStopRows(tripSnapshot.stopTimes);
+  const initialRenderStart = Math.max(
+    0,
+    selectedPatternStopIndex - INITIAL_TRIP_STOP_RENDER_PADDING,
+  );
+  const initialRenderEnd = Math.min(
+    renderedTripStopRows.length,
+    selectedPatternStopIndex + INITIAL_TRIP_STOP_RENDER_PADDING + 1,
+  );
+  const renderAllStops = renderedSnapshot === tripSnapshot;
+  const visibleTripStopRows = renderAllStops
+    ? renderedTripStopRows
+    : renderedTripStopRows.slice(initialRenderStart, initialRenderEnd);
+
+  return (
+    <section className="flex flex-col gap-2">
+      <div className="flex flex-col gap-2">
+        {visibleTripStopRows.map((row) => {
+          const rowKey =
+            row.kind === 'placeholder'
+              ? `placeholder:${row.stopIndex}`
+              : `${row.stop.stopMeta?.stop.stop_id || '(unknown-stop)'}:${row.stopIndex}`;
+
+          const rowContent =
+            row.kind === 'placeholder' ? (
+              <TripInspectionPlaceholderRow
+                stopIndex={row.stopIndex}
+                totalStops={row.totalStops}
+                currentPatternStopIndex={selectedPatternStopIndex}
+                infoLevel={infoLevel}
+              />
+            ) : (
+              <TripInspectionStopRow
+                tripStopTime={row.stop}
+                currentPatternStopIndex={selectedPatternStopIndex}
+                infoLevel={infoLevel}
+                dataLangs={dataLangs}
+                serviceDate={tripSnapshot.serviceDate}
+                now={now}
+              />
+            );
+
+          return (
+            <div key={rowKey} className="flex flex-col gap-1">
+              {infoLevelFlag.isVerboseEnabled && (
+                <div className="text-muted-foreground px-1 text-xs">
+                  {row.stopIndex + 1} / {row.totalStops}
+                </div>
+              )}
+              {rowContent}
+            </div>
+          );
+        })}
+      </div>
+    </section>
+  );
+}
